Create Setup the Frontend
'App.js': This file is the main component of the React.js application. It sets up the application's UI and state using React hooks like useState and useEffect.

Fetching Data: The use Effect hook is used to fetch tasks and users from the backend API when the component mounts.
Handling Form Input: The handle Change function updates the new taskk state as the user types into the form fields.

Submitting Form: The handleSubmit function is called when the form is submitted. It sends a POST request to the backend API to create a new task.

Displaying Data: The tasks and users fetched from the backend are displayed in the UI using map functions to iterate over arrays and render list items.